Data can now be saved
Data can now be saved to an excel file. We are using apache poi which
provides us with an api to read and write to excel files. The libraries
and dependencies are also included in this commit. Also an example of an
excel file that was outputted from our software is also included in this
commit